+++ deltachat-rpc-client/src/deltachat_rpc_client/chat.py
@@ -1,6 +1,7 @@
 import calendar
 from datetime import datetime
 from typing import TYPE_CHECKING, Dict, List, Optional, Tuple, Union
+from dataclasses import dataclass
 
 from ._utils import AttrDict
 from .const import ChatVisibility
@@ -12,27 +13,16 @@ if TYPE_CHECKING:
     from .account import Account
 
 
+@dataclass
 class Chat:
     """Chat object which manages members and through which you can send and retrieve messages."""
 
-    def __init__(self, account: "Account", chat_id: int) -> None:
-        self.account = account
-        self.id = chat_id
+    account: "Account"
+    id: int
 
     @property
     def _rpc(self) -> Rpc:
         return self.account._rpc
-
-    def __eq__(self, other) -> bool:
-        if not isinstance(other, Chat):
-            return False
-        return self.id == other.id and self.account == other.account
-
-    def __ne__(self, other) -> bool:
-        return not self == other
-
-    def __repr__(self) -> str:
-        return f"<Chat id={self.id} account={self.account.id}>"
 
     async def delete(self) -> None:
         """Delete this chat and all its messages.