+++ deltachat-rpc-client/src/deltachat_rpc_client/contact.py
@@ -1,4 +1,5 @@
 from typing import TYPE_CHECKING
+from dataclasses import dataclass
 
 from ._utils import AttrDict
 from .rpc import Rpc
@@ -8,6 +9,7 @@ if TYPE_CHECKING:
     from .chat import Chat
 
 
+@dataclass
 class Contact:
     """
     Contact API.
@@ -15,20 +17,8 @@ class Contact:
     Essentially a wrapper for RPC, account ID and a contact ID.
     """
 
-    def __init__(self, account: "Account", contact_id: int) -> None:
-        self.account = account
-        self.id = contact_id
-
-    def __eq__(self, other) -> bool:
-        if not isinstance(other, Contact):
-            return False
-        return self.id == other.id and self.account == other.account
-
-    def __ne__(self, other) -> bool:
-        return not self == other
-
-    def __repr__(self) -> str:
-        return f"<Contact id={self.id} account={self.account.id}>"
+    account: "Account"
+    id: int
 
     @property
     def _rpc(self) -> Rpc: