+++ deltachat-rpc-client/src/deltachat_rpc_client/message.py
@@ -1,5 +1,6 @@
 import json
 from typing import TYPE_CHECKING, Union
+from dataclasses import dataclass
 
 from ._utils import AttrDict
 from .contact import Contact
@@ -9,23 +10,12 @@ if TYPE_CHECKING:
     from .account import Account
 
 
+@dataclass
 class Message:
     """Delta Chat Message object."""
 
-    def __init__(self, account: "Account", msg_id: int) -> None:
-        self.account = account
-        self.id = msg_id
-
-    def __eq__(self, other) -> bool:
-        if not isinstance(other, Message):
-            return False
-        return self.id == other.id and self.account == other.account
-
-    def __ne__(self, other) -> bool:
-        return not self == other
-
-    def __repr__(self) -> str:
-        return f"<Message id={self.id} account={self.account.id}>"
+    account: "Account"
+    id: int
 
     @property
     def _rpc(self) -> Rpc: