+++ src/mimeparser.rs
@@ -48,13 +48,19 @@ use crate::{location, tools};
 /// It is created by parsing the raw data of an actual MIME message
 /// using the [MimeMessage::from_bytes] constructor.
 #[derive(Debug)]
-pub struct MimeMessage {
+pub(crate) struct MimeMessage {
+    /// Parsed MIME parts.
     pub parts: Vec<Part>,
+
+    /// Message headers.
     header: HashMap<String, String>,
 
     /// Addresses are normalized and lowercased:
     pub recipients: Vec<SingleInfo>,
+
+    /// `From:` address.
     pub from: SingleInfo,
+
     /// Whether the From address was repeated in the signed part
     /// (and we know that the signer intended to send from this address)
     pub from_is_signed: bool,
@@ -72,6 +78,8 @@ pub struct MimeMessage {
     /// The set of mail recipient addresses for which gossip headers were applied, regardless of
     /// whether they modified any peerstates.
     pub gossiped_addr: HashSet<String>,
+
+    /// True if the message is a forwarded message.
     pub is_forwarded: bool,
     pub is_system_message: SystemMessage,
     pub location_kml: Option<location::Kml>,
@@ -123,15 +131,26 @@ pub(crate) enum MailinglistType {
 }
 
 #[derive(
-    Debug, Display, Clone, Copy, PartialEq, Eq, FromPrimitive, ToPrimitive, ToSql, FromSql,
+    Debug, Default, Display, Clone, Copy, PartialEq, Eq, FromPrimitive, ToPrimitive, ToSql, FromSql,
 )]
 #[repr(u32)]
 pub enum SystemMessage {
+    #[default]
     Unknown = 0,
+
+    /// Group name changed.
     GroupNameChanged = 2,
+
+    /// Group avatar changed.
     GroupImageChanged = 3,
+
+    /// Member was added to the group.
     MemberAddedToGroup = 4,
+
+    /// Member was removed from the group.
     MemberRemovedFromGroup = 5,
+
+    /// Autocrypt Setup Message.
     AutocryptSetupMessage = 6,
     SecurejoinMessage = 7,
     LocationStreamingEnabled = 8,
@@ -140,41 +159,33 @@ pub enum SystemMessage {
     /// Chat ephemeral message timer is changed.
     EphemeralTimerChanged = 10,
 
-    // Chat protection state changed
+    /// Chat protection is enabled.
     ChatProtectionEnabled = 11,
+
+    /// Chat protection is disabled.
     ChatProtectionDisabled = 12,
 
     /// Self-sent-message that contains only json used for multi-device-sync;
     /// if possible, we attach that to other messages as for locations.
     MultiDeviceSync = 20,
 
-    // Sync message that contains a json payload
-    // sent to the other webxdc instances
-    // These messages are not shown in the chat.
+    /// Sync message that contains a json payload
+    /// sent to the other webxdc instances
+    /// These messages are not shown in the chat.
     WebxdcStatusUpdate = 30,
 
-    // Webxdc info added with `info` set in `send_webxdc_status_update()`.
+    /// Webxdc info added with `info` set in `send_webxdc_status_update()`.
     WebxdcInfoMessage = 32,
-}
-
-impl Default for SystemMessage {
-    fn default() -> Self {
-        SystemMessage::Unknown
-    }
 }
 
 const MIME_AC_SETUP_FILE: &str = "application/autocrypt-setup";
 
 impl MimeMessage {
-    pub async fn from_bytes(context: &Context, body: &[u8]) -> Result<Self> {
-        MimeMessage::from_bytes_with_partial(context, body, None).await
-    }
-
     /// Parse a mime message.
     ///
     /// If `partial` is set, it contains the full message size in bytes
     /// and `body` contains the header only.
-    pub(crate) async fn from_bytes_with_partial(
+    pub(crate) async fn from_bytes(
         context: &Context,
         body: &[u8],
         partial: Option<u32>,
@@ -1764,16 +1775,32 @@ fn is_known(key: &str) -> bool {
     )
 }
 
+/// Parsed MIME part.
 #[derive(Debug, Default, Clone)]
 pub struct Part {
+    /// Type of the MIME part determining how it should be displayed.
     pub typ: Viewtype,
+
+    /// MIME type.
     pub mimetype: Option<Mime>,
+
+    /// Message text to be displayed in the chat.
     pub msg: String,
+
+    /// Message text to be displayed in message info.
     pub msg_raw: Option<String>,
+
+    /// Size of the MIME part in bytes.
     pub bytes: usize,
     pub param: Params,
+
+    /// Attachment filename.
     pub(crate) org_filename: Option<String>,
+
+    /// An error detected during parsing.
     pub error: Option<String>,
+
+    /// True if conversion from HTML to plaintext failed.
     pub(crate) dehtml_failed: bool,
 
     /// the part is a child or a descendant of multipart/related.
@@ -1957,7 +1984,6 @@ mod tests {
     use super::*;
     use crate::{
         chatlist::Chatlist,
-        config::Config,
         constants::{Blocked, DC_DESIRED_TEXT_LEN, DC_ELLIPSIS},
         message::{Message, MessageState, MessengerMessage},
         receive_imf::receive_imf,
@@ -1977,35 +2003,35 @@ mod tests {
     async fn test_mimeparser_fromheader() {
         let ctx = TestContext::new_alice().await;
 
-        let mimemsg = MimeMessage::from_bytes(&ctx, b"From: g@c.de\n\nhi")
+        let mimemsg = MimeMessage::from_bytes(&ctx, b"From: g@c.de\n\nhi", None)
             .await
             .unwrap();
         let contact = mimemsg.from;
         assert_eq!(contact.addr, "g@c.de");
         assert_eq!(contact.display_name, None);
 
-        let mimemsg = MimeMessage::from_bytes(&ctx, b"From:   g@c.de  \n\nhi")
+        let mimemsg = MimeMessage::from_bytes(&ctx, b"From:   g@c.de  \n\nhi", None)
             .await
             .unwrap();
         let contact = mimemsg.from;
         assert_eq!(contact.addr, "g@c.de");
         assert_eq!(contact.display_name, None);
 
-        let mimemsg = MimeMessage::from_bytes(&ctx, b"From: <g@c.de>\n\nhi")
+        let mimemsg = MimeMessage::from_bytes(&ctx, b"From: <g@c.de>\n\nhi", None)
             .await
             .unwrap();
         let contact = mimemsg.from;
         assert_eq!(contact.addr, "g@c.de");
         assert_eq!(contact.display_name, None);
 
-        let mimemsg = MimeMessage::from_bytes(&ctx, b"From: Goetz C <g@c.de>\n\nhi")
+        let mimemsg = MimeMessage::from_bytes(&ctx, b"From: Goetz C <g@c.de>\n\nhi", None)
             .await
             .unwrap();
         let contact = mimemsg.from;
         assert_eq!(contact.addr, "g@c.de");
         assert_eq!(contact.display_name, Some("Goetz C".to_string()));
 
-        let mimemsg = MimeMessage::from_bytes(&ctx, b"From: \"Goetz C\" <g@c.de>\n\nhi")
+        let mimemsg = MimeMessage::from_bytes(&ctx, b"From: \"Goetz C\" <g@c.de>\n\nhi", None)
             .await
             .unwrap();
         let contact = mimemsg.from;
@@ -2013,7 +2039,7 @@ mod tests {
         assert_eq!(contact.display_name, Some("Goetz C".to_string()));
 
         let mimemsg =
-            MimeMessage::from_bytes(&ctx, b"From: =?utf-8?q?G=C3=B6tz?= C <g@c.de>\n\nhi")
+            MimeMessage::from_bytes(&ctx, b"From: =?utf-8?q?G=C3=B6tz?= C <g@c.de>\n\nhi", None)
                 .await
                 .unwrap();
         let contact = mimemsg.from;
@@ -2022,10 +2048,13 @@ mod tests {
 
         // although RFC 2047 says, encoded-words shall not appear inside quoted-string,
         // this combination is used in the wild eg. by MailMate
-        let mimemsg =
-            MimeMessage::from_bytes(&ctx, b"From: \"=?utf-8?q?G=C3=B6tz?= C\" <g@c.de>\n\nhi")
-                .await
-                .unwrap();
+        let mimemsg = MimeMessage::from_bytes(
+            &ctx,
+            b"From: \"=?utf-8?q?G=C3=B6tz?= C\" <g@c.de>\n\nhi",
+            None,
+        )
+        .await
+        .unwrap();
         let contact = mimemsg.from;
         assert_eq!(contact.addr, "g@c.de");
         assert_eq!(contact.display_name, Some("Götz C".to_string()));
@@ -2035,7 +2064,7 @@ mod tests {
     async fn test_mimeparser_crash() {
         let context = TestContext::new_alice().await;
         let raw = include_bytes!("../test-data/message/issue_523.txt");
-        let mimeparser = MimeMessage::from_bytes(&context.ctx, &raw[..])
+        let mimeparser = MimeMessage::from_bytes(&context.ctx, &raw[..], None)
             .await
             .unwrap();
 
@@ -2047,7 +2076,7 @@ mod tests {
     async fn test_get_rfc724_mid_exists() {
         let context = TestContext::new_alice().await;
         let raw = include_bytes!("../test-data/message/mail_with_message_id.txt");
-        let mimeparser = MimeMessage::from_bytes(&context.ctx, &raw[..])
+        let mimeparser = MimeMessage::from_bytes(&context.ctx, &raw[..], None)
             .await
             .unwrap();
 
@@ -2061,7 +2090,7 @@ mod tests {
     async fn test_get_rfc724_mid_not_exists() {
         let context = TestContext::new_alice().await;
         let raw = include_bytes!("../test-data/message/issue_523.txt");
-        let mimeparser = MimeMessage::from_bytes(&context.ctx, &raw[..])
+        let mimeparser = MimeMessage::from_bytes(&context.ctx, &raw[..], None)
             .await
             .unwrap();
         assert_eq!(mimeparser.get_rfc724_mid(), None);
@@ -2258,7 +2287,7 @@ mod tests {
                     test1\n\
                     ";
 
-        let mimeparser = MimeMessage::from_bytes_with_partial(&context.ctx, &raw[..], None).await;
+        let mimeparser = MimeMessage::from_bytes(&context.ctx, &raw[..], None).await;
 
         assert!(mimeparser.is_err());
     }
@@ -2273,7 +2302,7 @@ mod tests {
                     \n\
                     Some reply\n\
                     ";
-        let mimeparser = MimeMessage::from_bytes(&context.ctx, &raw[..])
+        let mimeparser = MimeMessage::from_bytes(&context.ctx, &raw[..], None)
             .await
             .unwrap();
         assert_eq!(
@@ -2319,7 +2348,7 @@ mod tests {
                     --==break==--\n\
                     \n";
 
-        let mimeparser = MimeMessage::from_bytes(&context.ctx, &raw[..])
+        let mimeparser = MimeMessage::from_bytes(&context.ctx, &raw[..], None)
             .await
             .unwrap();
 
@@ -2328,7 +2357,7 @@ mod tests {
         assert_eq!(of, "no");
 
         // unknown headers do not bubble upwards
-        let of = mimeparser.get_header(HeaderDef::_TestHeader).unwrap();
+        let of = mimeparser.get_header(HeaderDef::TestHeader).unwrap();
         assert_eq!(of, "Bar");
 
         // the following fields would bubble up
@@ -2353,26 +2382,26 @@ mod tests {
         let t = TestContext::new_alice().await;
 
         let raw = include_bytes!("../test-data/message/mail_attach_txt.eml");
-        let mimeparser = MimeMessage::from_bytes(&t, &raw[..]).await.unwrap();
+        let mimeparser = MimeMessage::from_bytes(&t, &raw[..], None).await.unwrap();
         assert_eq!(mimeparser.user_avatar, None);
         assert_eq!(mimeparser.group_avatar, None);
 
         let raw = include_bytes!("../test-data/message/mail_with_user_avatar.eml");
-        let mimeparser = MimeMessage::from_bytes(&t, &raw[..]).await.unwrap();
+        let mimeparser = MimeMessage::from_bytes(&t, &raw[..], None).await.unwrap();
         assert_eq!(mimeparser.parts.len(), 1);
         assert_eq!(mimeparser.parts[0].typ, Viewtype::Text);
         assert!(mimeparser.user_avatar.unwrap().is_change());
         assert_eq!(mimeparser.group_avatar, None);
 
         let raw = include_bytes!("../test-data/message/mail_with_user_avatar_deleted.eml");
-        let mimeparser = MimeMessage::from_bytes(&t, &raw[..]).await.unwrap();
+        let mimeparser = MimeMessage::from_bytes(&t, &raw[..], None).await.unwrap();
         assert_eq!(mimeparser.parts.len(), 1);
         assert_eq!(mimeparser.parts[0].typ, Viewtype::Text);
         assert_eq!(mimeparser.user_avatar, Some(AvatarAction::Delete));
         assert_eq!(mimeparser.group_avatar, None);
 
         let raw = include_bytes!("../test-data/message/mail_with_user_and_group_avatars.eml");
-        let mimeparser = MimeMessage::from_bytes(&t, &raw[..]).await.unwrap();
+        let mimeparser = MimeMessage::from_bytes(&t, &raw[..], None).await.unwrap();
         assert_eq!(mimeparser.parts.len(), 1);
         assert_eq!(mimeparser.parts[0].typ, Viewtype::Text);
         assert!(mimeparser.user_avatar.unwrap().is_change());
@@ -2382,7 +2411,9 @@ mod tests {
         let raw = include_bytes!("../test-data/message/mail_with_user_and_group_avatars.eml");
         let raw = String::from_utf8_lossy(raw).to_string();
         let raw = raw.replace("Chat-User-Avatar:", "Xhat-Xser-Xvatar:");
-        let mimeparser = MimeMessage::from_bytes(&t, raw.as_bytes()).await.unwrap();
+        let mimeparser = MimeMessage::from_bytes(&t, raw.as_bytes(), None)
+            .await
+            .unwrap();
         assert_eq!(mimeparser.parts.len(), 1);
         assert_eq!(mimeparser.parts[0].typ, Viewtype::Image);
         assert_eq!(mimeparser.user_avatar, None);
@@ -2394,7 +2425,7 @@ mod tests {
         let t = TestContext::new_alice().await;
 
         let raw = include_bytes!("../test-data/message/videochat_invitation.eml");
-        let mimeparser = MimeMessage::from_bytes(&t, &raw[..]).await.unwrap();
+        let mimeparser = MimeMessage::from_bytes(&t, &raw[..], None).await.unwrap();
         assert_eq!(mimeparser.parts.len(), 1);
         assert_eq!(mimeparser.parts[0].typ, Viewtype::VideochatInvitation);
         assert_eq!(
@@ -2441,7 +2472,7 @@ Content-Disposition: attachment; filename=\"message.kml\"\n\
 --==break==--\n\
 ;";
 
-        let mimeparser = MimeMessage::from_bytes(&context.ctx, &raw[..])
+        let mimeparser = MimeMessage::from_bytes(&context.ctx, &raw[..], None)
             .await
             .unwrap();
         assert_eq!(
@@ -2490,7 +2521,7 @@ Disposition: manual-action/MDN-sent-automatically; displayed\n\
 --kJBbU58X1xeWNHgBtTbMk80M5qnV4N--\n\
 ";
 
-        let message = MimeMessage::from_bytes(&context.ctx, &raw[..])
+        let message = MimeMessage::from_bytes(&context.ctx, &raw[..], None)
             .await
             .unwrap();
         assert_eq!(
@@ -2570,7 +2601,7 @@ Disposition: manual-action/MDN-sent-automatically; displayed\n\
 --outer--\n\
 ";
 
-        let message = MimeMessage::from_bytes(&context.ctx, &raw[..])
+        let message = MimeMessage::from_bytes(&context.ctx, &raw[..], None)
             .await
             .unwrap();
         assert_eq!(
@@ -2617,7 +2648,7 @@ Additional-Message-IDs: <foo@example.com> <foo@example.net>\n\
 --kJBbU58X1xeWNHgBtTbMk80M5qnV4N--\n\
 ";
 
-        let message = MimeMessage::from_bytes(&context.ctx, &raw[..])
+        let message = MimeMessage::from_bytes(&context.ctx, &raw[..], None)
             .await
             .unwrap();
         assert_eq!(
@@ -2664,7 +2695,7 @@ MDYyMDYxNTE1RTlDOEE4Cj4+CnN0YXJ0eHJlZgo4Mjc4CiUlRU9GCg==
 ------=_Part_25_46172632.1581201680436--
 "#;
 
-        let message = MimeMessage::from_bytes(&context.ctx, &raw[..])
+        let message = MimeMessage::from_bytes(&context.ctx, &raw[..], None)
             .await
             .unwrap();
         assert_eq!(
@@ -2708,7 +2739,7 @@ MDYyMDYxNTE1RTlDOEE4Cj4+CnN0YXJ0eHJlZgo4Mjc4CiUlRU9GCg==
 ------=_Part_25_46172632.1581201680436--
 "#;
 
-        let message = MimeMessage::from_bytes(&t, &raw[..]).await.unwrap();
+        let message = MimeMessage::from_bytes(&t, &raw[..], None).await.unwrap();
 
         assert_eq!(message.parts.len(), 1);
         assert_eq!(message.parts[0].typ, Viewtype::File);
@@ -2762,7 +2793,7 @@ CWt6wx7fiLp0qS9RrX75g6Gqw7nfCs6EcBERcIPt7DTe8VStJwf3LWqVwxl4gQl46yhfoqwEO+I=
 ----11019878869865180--
 "#;
 
-        let message = MimeMessage::from_bytes(&context.ctx, &raw[..])
+        let message = MimeMessage::from_bytes(&context.ctx, &raw[..], None)
             .await
             .unwrap();
         assert_eq!(message.get_subject(), Some("example".to_string()));
@@ -2834,7 +2865,7 @@ CWt6wx7fiLp0qS9RrX75g6Gqw7nfCs6EcBERcIPt7DTe8VStJwf3LWqVwxl4gQl46yhfoqwEO+I=
 
 --------------779C1631600DF3DB8C02E53A--"#;
 
-        let message = MimeMessage::from_bytes(&context.ctx, &raw[..])
+        let message = MimeMessage::from_bytes(&context.ctx, &raw[..], None)
             .await
             .unwrap();
         assert_eq!(message.get_subject(), Some("Test subject".to_string()));
@@ -2905,7 +2936,7 @@ CWt6wx7fiLp0qS9RrX75g6Gqw7nfCs6EcBERcIPt7DTe8VStJwf3LWqVwxl4gQl46yhfoqwEO+I=
 ------=_NextPart_000_0003_01D622B3.CA753E60--
 "##;
 
-        let message = MimeMessage::from_bytes(&context.ctx, &raw[..])
+        let message = MimeMessage::from_bytes(&context.ctx, &raw[..], None)
             .await
             .unwrap();
         assert_eq!(
@@ -3003,7 +3034,7 @@ From: alice <alice@example.org>
 Reply
 "##;
 
-        let message = MimeMessage::from_bytes(&context.ctx, &raw[..])
+        let message = MimeMessage::from_bytes(&context.ctx, &raw[..], None)
             .await
             .unwrap();
         assert_eq!(
@@ -3035,7 +3066,7 @@ From: alice <alice@example.org>
 > Just a quote.
 "##;
 
-        let message = MimeMessage::from_bytes(&context.ctx, &raw[..])
+        let message = MimeMessage::from_bytes(&context.ctx, &raw[..], None)
             .await
             .unwrap();
         assert_eq!(
@@ -3069,7 +3100,7 @@ On 2020-10-25, Bob wrote:
 > A quote.
 "##;
 
-        let message = MimeMessage::from_bytes(&context.ctx, &raw[..])
+        let message = MimeMessage::from_bytes(&context.ctx, &raw[..], None)
             .await
             .unwrap();
         assert_eq!(message.get_subject(), Some("Re: top posting".to_string()));
@@ -3087,7 +3118,7 @@ On 2020-10-25, Bob wrote:
     async fn test_attachment_quote() {
         let context = TestContext::new_alice().await;
         let raw = include_bytes!("../test-data/message/quote_attach.eml");
-        let mimeparser = MimeMessage::from_bytes(&context.ctx, &raw[..])
+        let mimeparser = MimeMessage::from_bytes(&context.ctx, &raw[..], None)
             .await
             .unwrap();
 
@@ -3105,7 +3136,7 @@ On 2020-10-25, Bob wrote:
     async fn test_quote_div() {
         let t = TestContext::new_alice().await;
         let raw = include_bytes!("../test-data/message/gmx-quote.eml");
-        let mimeparser = MimeMessage::from_bytes(&t, raw).await.unwrap();
+        let mimeparser = MimeMessage::from_bytes(&t, raw, None).await.unwrap();
         assert_eq!(mimeparser.parts[0].msg, "YIPPEEEEEE\n\nMulti-line");
         assert_eq!(mimeparser.parts[0].param.get(Param::Quote).unwrap(), "Now?");
     }
@@ -3115,7 +3146,7 @@ On 2020-10-25, Bob wrote:
         // all-inkl.com puts quotes into `<blockquote> </blockquote>`.
         let t = TestContext::new_alice().await;
         let raw = include_bytes!("../test-data/message/allinkl-quote.eml");
-        let mimeparser = MimeMessage::from_bytes(&t, raw).await.unwrap();
+        let mimeparser = MimeMessage::from_bytes(&t, raw, None).await.unwrap();
         assert!(mimeparser.parts[0].msg.starts_with("It's 1.0."));
         assert_eq!(
             mimeparser.parts[0].param.get(Param::Quote).unwrap(),
@@ -3126,7 +3157,6 @@ On 2020-10-25, Bob wrote:
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
     async fn test_add_subj_to_multimedia_msg() {
         let t = TestContext::new_alice().await;
-        t.set_config(Config::ShowEmails, Some("2")).await.unwrap();
         receive_imf(
             &t.ctx,
             include_bytes!("../test-data/message/subj_with_multimedia_msg.eml"),
@@ -3160,7 +3190,7 @@ On 2020-10-25, Bob wrote:
     async fn test_mime_modified_plain() {
         let t = TestContext::new_alice().await;
         let raw = include_bytes!("../test-data/message/text_plain_unspecified.eml");
-        let mimeparser = MimeMessage::from_bytes(&t.ctx, raw).await.unwrap();
+        let mimeparser = MimeMessage::from_bytes(&t.ctx, raw, None).await.unwrap();
         assert!(!mimeparser.is_mime_modified);
         assert_eq!(
             mimeparser.parts[0].msg,
@@ -3172,7 +3202,7 @@ On 2020-10-25, Bob wrote:
     async fn test_mime_modified_alt_plain_html() {
         let t = TestContext::new_alice().await;
         let raw = include_bytes!("../test-data/message/text_alt_plain_html.eml");
-        let mimeparser = MimeMessage::from_bytes(&t.ctx, raw).await.unwrap();
+        let mimeparser = MimeMessage::from_bytes(&t.ctx, raw, None).await.unwrap();
         assert!(mimeparser.is_mime_modified);
         assert_eq!(
             mimeparser.parts[0].msg,
@@ -3184,7 +3214,7 @@ On 2020-10-25, Bob wrote:
     async fn test_mime_modified_alt_plain() {
         let t = TestContext::new_alice().await;
         let raw = include_bytes!("../test-data/message/text_alt_plain.eml");
-        let mimeparser = MimeMessage::from_bytes(&t.ctx, raw).await.unwrap();
+        let mimeparser = MimeMessage::from_bytes(&t.ctx, raw, None).await.unwrap();
         assert!(!mimeparser.is_mime_modified);
         assert_eq!(
             mimeparser.parts[0].msg,
@@ -3199,7 +3229,7 @@ On 2020-10-25, Bob wrote:
     async fn test_mime_modified_alt_html() {
         let t = TestContext::new_alice().await;
         let raw = include_bytes!("../test-data/message/text_alt_html.eml");
-        let mimeparser = MimeMessage::from_bytes(&t.ctx, raw).await.unwrap();
+        let mimeparser = MimeMessage::from_bytes(&t.ctx, raw, None).await.unwrap();
         assert!(mimeparser.is_mime_modified);
         assert_eq!(
             mimeparser.parts[0].msg,
@@ -3211,7 +3241,7 @@ On 2020-10-25, Bob wrote:
     async fn test_mime_modified_html() {
         let t = TestContext::new_alice().await;
         let raw = include_bytes!("../test-data/message/text_html.eml");
-        let mimeparser = MimeMessage::from_bytes(&t.ctx, raw).await.unwrap();
+        let mimeparser = MimeMessage::from_bytes(&t.ctx, raw, None).await.unwrap();
         assert!(mimeparser.is_mime_modified);
         assert_eq!(
             mimeparser.parts[0].msg,
@@ -3227,7 +3257,7 @@ On 2020-10-25, Bob wrote:
         static REPEAT_CNT: usize = 2000; // results in a text of 84k, should be more than DC_DESIRED_TEXT_LEN
         let long_txt = format!("From: alice@c.de\n\n{}", REPEAT_TXT.repeat(REPEAT_CNT));
 
-        let mimemsg = MimeMessage::from_bytes(&t, long_txt.as_ref())
+        let mimemsg = MimeMessage::from_bytes(&t, long_txt.as_ref(), None)
             .await
             .unwrap();
         assert_eq!(long_txt.matches("just repeated").count(), REPEAT_CNT);
@@ -3254,7 +3284,7 @@ On 2020-10-25, Bob wrote:
                 MIME-Version: 1.0\n\
                 \n\
                 Does it work with outlook now?\n\
-                ")
+                ", None)
             .await
             .unwrap();
         assert_eq!(
@@ -3410,7 +3440,6 @@ Message.
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
     async fn test_ms_exchange_mdn() -> Result<()> {
         let t = TestContext::new_alice().await;
-        t.set_config(Config::ShowEmails, Some("2")).await?;
 
         let original =
             include_bytes!("../test-data/message/ms_exchange_report_original_message.eml");
@@ -3420,7 +3449,7 @@ Message.
         // 1. Test mimeparser directly
         let mdn =
             include_bytes!("../test-data/message/ms_exchange_report_disposition_notification.eml");
-        let mimeparser = MimeMessage::from_bytes(&t.ctx, mdn).await?;
+        let mimeparser = MimeMessage::from_bytes(&t.ctx, mdn, None).await?;
         assert_eq!(mimeparser.mdn_reports.len(), 1);
         assert_eq!(
             mimeparser.mdn_reports[0].original_message_id.as_deref(),
@@ -3446,6 +3475,7 @@ Message.
         let mime_message = MimeMessage::from_bytes(
             &alice,
             include_bytes!("../test-data/message/attached-eml.eml"),
+            None,
         )
         .await?;
 
@@ -3488,6 +3518,7 @@ Content-Disposition: reaction\n\
 \n\
 \u{1F44D}"
                 .as_bytes(),
+            None,
         )
         .await?;